# protocol/openapi.yaml
openapi: 3.1.0
info:
  title: CapOwn Control Plane API
  version: 1.6.0
  description: |
    Language-independent v1.6 wire contract for the CapOwn control plane.

    Covers the connectivity milestone, plugin discovery, task dispatch,
    claim-based job delivery, results, cancellation, and plugin invocation
    via the `plugin_call` task type. Workers claim jobs via long-poll;
    SSE is an optional wake accelerator.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  x-capown-protocol-version: "1.6"
  x-capown-plugin-error-codes:
    - plugin_not_found
    - plugin_disabled
    - plugin_unavailable
    - plugin_tool_not_found
    - plugin_schema_invalid
    - plugin_timeout
    - plugin_canceled
    - plugin_output_too_large
    - plugin_protocol_error
    - plugin_concurrency_exceeded
  x-capown-task-error-codes:
    - worker_offline
  x-capown-auth-error-codes:
    - user_disabled
    - invitation_invalid
servers:
  - url: http://localhost:9210
    description: Local development Master
...
paths:
...
  /v1/auth/register:
    post:
      tags: [auth]
      summary: Register a Master user
      description: |
        The first user becomes an administrator and does not provide an
        invitation code. Once initialized, registration creates a normal user
        and requires a valid, unused administrator invitation code.
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        "201":
          description: User created and a web session issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        "400":
          $ref: '#/components/responses/InvalidInput'
        "409":
          $ref: '#/components/responses/Conflict'
  /v1/admin/invitations:
    get:
      tags: [admin, invitations]
      summary: List user invitations
      operationId: adminListInvitations
      security:
        - bearerAuth: []
      x-capown-auth-token-types: [web, admin]
      responses:
        "200":
          description: Invitation list without plaintext codes.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Invitation'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [admin, invitations]
      summary: Create a one-time user invitation
      operationId: adminCreateInvitation
      security:
        - bearerAuth: []
      x-capown-auth-token-types: [web, admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvitationRequest'
      responses:
        "201":
          description: Invitation created. The plaintext code is returned only here.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedInvitation'
        "400":
          $ref: '#/components/responses/InvalidInput'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
  /v1/admin/invitations/{invitation_id}:
    delete:
      tags: [admin, invitations]
      summary: Revoke a user invitation
      operationId: adminRevokeInvitation
      security:
        - bearerAuth: []
      x-capown-auth-token-types: [web, admin]
      parameters:
        - $ref: '#/components/parameters/InvitationId'
      responses:
        "204":
          description: Invitation revoked.
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
  /v1/auth/login:
    post:
      tags: [auth]
      summary: Create a web session
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialsRequest'
      responses:
        "200":
          description: Web session created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        "400":
          $ref: '#/components/responses/InvalidInput'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
  /v1/auth/logout:
    post:
      tags: [auth]
      summary: Revoke the current web session
      operationId: logout
      security:
        - bearerAuth: []
      x-capown-auth-token-types: [web]
      responses:
        "200":
          description: Session revoked.
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    const: logged_out
        "401":
          $ref: '#/components/responses/Unauthorized'
  /v1/me:
    get:
      tags: [auth]
      summary: Get the current user
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      x-capown-auth-token-types: [web]
      responses:
        "200":
          description: Current user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /v1/me/password:
    patch:
      tags: [auth]
      summary: Change the current user's password
      operationId: changePassword
      security:
        - bearerAuth: []
      x-capown-auth-token-types: [web]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        "200":
          description: Password changed.
          content:
            application/json:
              schema:
                type: object
                required: [status, sessions_revoked]
                properties:
                  status:
                    type: string
                    const: password_updated
                  sessions_revoked:
                    type: integer
                    minimum: 0
        "400":
          $ref: '#/components/responses/InvalidInput'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /v1/tokens:
    get:
      tags: [tokens]
      summary: List tokens owned by the current user
      operationId: listTokens
      security:
        - bearerAuth: []
      x-capown-auth-token-types: [web]
      parameters:
        - $ref: '#/components/parameters/TokenTypeQuery'
      responses:
        "200":
          description: Token list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenListResponse'
        "401":
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [tokens]
      summary: Create a client or admin token
      operationId: createToken
      security:
        - bearerAuth: []
      x-capown-auth-token-types: [web]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTokenRequest'
      responses:
        "201":
          description: Token created. The plaintext token is returned only here.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedToken'
        "400":
          $ref: '#/components/responses/InvalidInput'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
  /v1/tokens/{token_id}:
    delete:
      tags: [tokens]
      summary: Revoke an owned token
      operationId: revokeToken
      security:
        - bearerAuth: []
      x-capown-auth-token-types: [web]
      parameters:
        - $ref: '#/components/parameters/TokenId'
      responses:
        "204":
          description: Token revoked.
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
  /v1/worker-registrations:
    get:
      tags: [worker-registrations]
      summary: List Worker registration tokens
      operationId: listWorkerRegistrations
      security:
        - bearerAuth: []
      x-capown-auth-token-types: [web]
      responses:
        "200":
          description: Registration token list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RegistrationTokenView'
        "401":
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [worker-registrations]
      summary: Create a Worker registration token
      operationId: createWorkerRegistration
      security:
        - bearerAuth: []
      x-capown-auth-token-types: [web]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRegistrationRequest'
      responses:
        "201":
          description: Registration token created. The plaintext token and registration URL are returned only here.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedRegistrationToken'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /v1/worker-registrations/{registration_id}:
    delete:
      tags: [worker-registrations]
      summary: Revoke an owned Worker registration token
      operationId: revokeWorkerRegistration
      security:
        - bearerAuth: []
      x-capown-auth-token-types: [web]
      parameters:
        - $ref: '#/components/parameters/RegistrationId'
      responses:
        "204":
          description: Registration token revoked.
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
  /v1/workers:
    get:
      tags: [workers]
      summary: List active Workers visible to the caller
      operationId: listWorkers
      security:
        - bearerAuth: []
      x-capown-auth-token-types: [web, client, admin, worker]
      responses:
        "200":
          description: Worker list ordered by registration time.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkerListResponse'
        "401":
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [workers]
      summary: Register a Worker using a registration token
      operationId: registerWorker
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkerRegistrationRequest'
      responses:
        "201":
          description: Worker registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkerRegistrationResponse'
        "400":
          $ref: '#/components/responses/InvalidInput'
        "401":
          $ref: '#/components/responses/RegistrationUnauthorized'
        "409":
          $ref: '#/components/responses/RegistrationConflict'
  /v1/workers/auth/challenges:
    post:
      tags: [workers]
      summary: Issue an Ed25519 authentication challenge
      operationId: createWorkerChallenge
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkerAuthChallengeRequest'
      responses:
        "200":
          description: Challenge issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkerAuthChallengeResponse'
        "400":
          $ref: '#/components/responses/InvalidInput'
        "404":
          $ref: '#/components/responses/WorkerNotFound'
  /v1/workers/auth/sessions:
    post:
      tags: [workers]
      summary: Verify a signed challenge and create a Worker session
      operationId: createWorkerSession
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkerAuthSessionRequest'
      responses:
        "200":
          description: Worker session created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkerAuthSessionResponse'
        "400":
          $ref: '#/components/responses/InvalidInput'
        "401":
          $ref: '#/components/responses/SignatureUnauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
  /v1/workers/{worker_id}:
    parameters:
      - $ref: '#/components/parameters/WorkerId'
    get:
      tags: [workers]
      summary: Get one visible Worker
      operationId: getWorker
      security:
        - bearerAuth: []
      x-capown-auth-token-types: [web, client, admin, worker]
      responses:
        "200":
          description: Worker information.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Worker'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/WorkerNotFound'
    patch:
      tags: [workers]
      summary: Rename a Worker
      operationId: renameWorker
      security:
        - bearerAuth: []
      x-capown-auth-token-types: [web]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenameWorkerRequest'
      responses:
        "200":
          description: Worker renamed.
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    const: updated
        "400":
          $ref: '#/components/responses/InvalidInput'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/WorkerNotFound'
        "409":
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [workers]
      summary: Revoke a Worker
      operationId: revokeWorker
      security:
        - bearerAuth: []
      x-capown-auth-token-types: [web]
      responses:
        "204":
          description: Worker revoked.
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/WorkerNotFound'
  /v1/workers/{worker_id}/runtime:
    put:
      tags: [workers]
      summary: Report Worker runtime metadata
      operationId: updateWorkerRuntime
      security:
        - bearerAuth: []
      x-capown-auth-token-types: [worker]
      parameters:
        - $ref: '#/components/parameters/WorkerId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkerRuntimeUpdateRequest'
      responses:
        "200":
          description: Runtime metadata accepted.
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    const: ok
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/WorkerNotFound'
  /v1/workers/{worker_id}/plugins:
    get:
      tags: [plugins, workers]
      summary: List plugins for a Worker
      operationId: listWorkerPlugins
      security:
        - bearerAuth: []
      x-capown-auth-token-types: [web, client, admin, worker]
      parameters:
        - $ref: '#/components/parameters/WorkerId'
      responses:
        "200":
          description: Plugin list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PluginInfo'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/WorkerNotFound'
  /v1/workers/{worker_id}/plugins/{plugin_id}:
    patch:
      tags: [plugins, workers]
      summary: Enable or disable a Worker plugin
      description: |
        Persistently changes the local plugin manifest through the Worker task
        channel. The Worker must be online. This endpoint cannot change plugin
        commands, environment variables, permissions, or installation state.
      operationId: setWorkerPluginEnabled
      security:
        - bearerAuth: []
      x-capown-auth-token-types: [web]
      parameters:
        - $ref: '#/components/parameters/WorkerId'
        - $ref: '#/components/parameters/PluginId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetPluginEnabledRequest'
      responses:
        "202":
          description: Plugin state change task accepted.
          headers:
            Location:
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        "400":
          $ref: '#/components/responses/TaskRequestRejected'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
  /v1/workers/{worker_id}/events:
    get:
      tags: [events, workers]
      summary: Open the optional Worker wake SSE stream
      description: |
        This stream is authenticated with a Worker session and is consumed by
        the Worker. It is an optional accelerator only: the Master may emit
        `wake` events when jobs are available so the Worker can claim sooner.
        Task payloads are never delivered on this stream. Correctness depends
        on `POST /v1/workers/{worker_id}/jobs/claim`, not on SSE connectivity.
        Heartbeat comments (`: ping`) are sent at regular intervals.
      operationId: workerEventStream
      security:
        - bearerAuth: []
      x-capown-auth-token-types: [worker]
      parameters:
        - $ref: '#/components/parameters/WorkerId'
      responses:
        "200":
          description: SSE stream.
          content:
            text/event-stream:
              schema:
                type: string
                description: SSE wire stream. See components.schemas.SSEEnvelope.
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/WorkerNotFound'
  /v1/workers/{worker_id}/jobs/claim:
    post:
      tags: [tasks, workers]
      summary: Claim pending jobs for this Worker (long-poll)
      description: |
        Reserve pending work and cancel jobs for delivery to the authenticated
        Worker. This endpoint is the source of truth for task delivery.

        When `wait_seconds` is greater than zero and no jobs are immediately
        available, the Master holds the request until jobs arrive or the wait
        expires. An empty `jobs` array is returned on timeout.

        A returned `task` job has a short internal delivery lease. The task
        remains `pending` until the Worker confirms receipt by reporting
        `status: running` with the job's opaque `delivery_id`. The Worker MUST
        confirm receipt before execution and include the same `delivery_id` in
        every later result for that execution.
        If confirmation does not arrive before the lease expires, the Master
        makes the job claimable again.

        A returned `cancel` job is also leased and may be delivered again until
        the Worker reports a terminal task result. Workers MUST therefore treat
        task and cancel delivery as idempotent. The optional SSE `wake` event
        MUST NOT cancel an active claim request.

        A Worker is considered online for dispatch while it has an open SSE
        stream or an active long-poll claim request.
      operationId: claimWorkerJobs
      security:
        - bearerAuth: []
      x-capown-auth-token-types: [worker]
      parameters:
        - $ref: '#/components/parameters/WorkerId'
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 10
            default: 1
          description: Maximum number of jobs to claim.
        - name: wait_seconds
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            maximum: 30
            default: 25
          description: |
            Long-poll wait in seconds. `0` returns immediately with whatever
            jobs are available (possibly empty).
      responses:
        "200":
          description: Claimed jobs, or an empty list when none were available.
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
            Vary:
              schema:
                type: string
                const: Authorization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkerJobsResponse'
        "400":
          $ref: '#/components/responses/InvalidInput'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/WorkerNotFound'
  /v1/events:
    get:
      tags: [events]
      summary: Open the dashboard event SSE stream
      description: |
        Dashboard events include an SSE id and can be replayed by sending the
        Last-Event-ID request header. A resync event is emitted when the id is
        no longer available in the in-memory history. Normal users receive
        events for their own resources. Administrators receive the global
        Worker event stream for every user.
      operationId: dashboardEventStream
      security:
        - bearerAuth: []
      x-capown-auth-token-types: [web]
      parameters:
        - name: Last-Event-ID
          in: header
          required: false
          schema:
            type: string
          description: Last dashboard event id received by the client.
      responses:
        "200":
          description: Dashboard SSE stream.
          content:
            text/event-stream:
              schema:
                type: string
                description: SSE wire stream. See components.schemas.DashboardSSEEnvelope.
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
...
  /v1/tasks/{task_id}:
    get:
      tags: [tasks]
      summary: Get task status and result
      operationId: getTask
      security:
        - bearerAuth: []
      x-capown-auth-token-types: [web, client, admin]
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        "200":
          description: Task status and result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
  /v1/tasks/{task_id}/result:
    put:
      tags: [tasks]
      summary: Report task result (Worker-to-Master only)
      operationId: reportTaskResult
      security:
        - bearerAuth: []
      x-capown-auth-token-types: [worker]
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskResult'
      responses:
        "200":
          description: Result accepted.
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    const: ok
        "400":
          $ref: '#/components/responses/InvalidInput'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Send `Authorization: Bearer <token>`. Token classes are identified by
        server-side storage and, for web/worker sessions, by the cown_* prefix.
...
  schemas:
...
    RegisterRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 6
        invitation_code:
          type: string
          description: Required after the first administrator initializes the Master.
    LoginResponse:
      type: object
      additionalProperties: false
      required: [access_token, token_type, expires_at, user]
      properties:
        access_token:
          type: string
          description: Opaque web session token. Treat as a secret.
        token_type:
          type: string
          const: bearer
        expires_at:
          $ref: '#/components/schemas/Timestamp'
        user:
          $ref: '#/components/schemas/User'
...